Refactor code and implement bulk CRUD
