+++ company/views.py
@@ -15,7 +15,7 @@ from core.models import (
     User,
     PaymentMethod,
 )
-from core.utils import validate_reference_uniqueness_in_data
+from core.utils import validate_bulk_reference_uniqueness
 from company import serializers
 from user.serializers import OwnerProfileSerializer
 
@@ -92,13 +92,13 @@ class ProductViewSet(BaseAttrViewSet):
         company = self.request.user.company
         return Product.objects.filter(category__company=company)
 
-    def perform_create(self, serializer):
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        serializer.save()
+    def perform_bulk_create(self, serializer):
+        validate_bulk_reference_uniqueness(serializer.validated_data)
+        return self.perform_create(serializer)
 
-    def perform_update(self, serializer):
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        serializer.save()
+    def perform_bulk_update(self, serializer):
+        validate_bulk_reference_uniqueness(serializer.validated_data)
+        return self.perform_update(serializer)
 
 
 class PaymentMethodViewSet(BaseAssetAttrViewSet):
