+++ customer/serializers.py
@@ -1,7 +1,9 @@
-from django.db.models.query import QuerySet
+from decimal import Decimal
+from datetime import datetime
+
+from django.utils import formats
 from django.utils.translation import ugettext_lazy as _
 from django.contrib.auth import get_user_model
-
 
 from rest_framework import serializers
 from rest_framework_bulk import (
@@ -249,7 +251,67 @@ class CreditNoteSerializer(DocumentSerializer):
         ).delete()
         CreditNoteItem.objects.bulk_create(bulk_creates)
 
+    def _get_calculated_fields(self, validated_data):
+        # assume all positive
+        discount_rate = validated_data.pop("discount_rate")
+        gst_rate = validated_data.pop("gst_rate")
+        creditnoteitem_set = validated_data.pop("creditnoteitem_set")
+        credits_used = Decimal("0.00")
+        # TODO: remove this later after implement apply credits to invoice from credit note
+        if self.context["request"].method in ["PUT", "PATCH"]:
+            credits_used = self.instance.credits_used
+
+        refund = validated_data.pop("refund")
+
+        creditnoteitem_set = [
+            {
+                **creditnoteitem,
+                "amount": round(
+                    round(creditnoteitem.get("quantity"))
+                    * round(creditnoteitem.get("unit_price"), 2),
+                    2,
+                ),
+            }
+            for creditnoteitem in creditnoteitem_set
+        ]
+        total_amount = sum(
+            [
+                # recalculate amount here since it has been rounded up
+                round(creditnoteitem.get("quantity"))
+                * round(creditnoteitem.get("unit_price"), 2)
+                for creditnoteitem in creditnoteitem_set
+            ]
+        )
+        discount_rate = round(discount_rate, 2)
+        gst_rate = round(gst_rate, 2)
+        discount_amount = total_amount * discount_rate / 100
+        net = total_amount * (1 - discount_rate / 100)
+        gst_amount = net * gst_rate / 100
+        grand_total = net * (1 + gst_rate / 100)
+        credits_remaining = grand_total - credits_used
+        refund = min(credits_remaining, round(refund, 2))
+        credits_remaining -= refund
+
+        return {
+            "total_amount": round(total_amount, 2),
+            "discount_rate": discount_rate,
+            "gst_rate": gst_rate,
+            "discount_amount": round(discount_amount, 2),
+            "gst_amount": round(gst_amount, 2),
+            "net": round(net, 2),
+            "grand_total": round(grand_total, 2),
+            "creditnoteitem_set": creditnoteitem_set,
+            "credits_used": credits_used,
+            "credits_remaining": credits_remaining,
+            "refund": refund,
+        }
+
     def create(self, validated_data):
+        validated_data = {
+            **validated_data,
+            **self._get_calculated_fields(validated_data),
+        }
+
         creditnoteitems_data = validated_data.pop("creditnoteitem_set", [])
         credit_note = CreditNote.objects.create(**validated_data)
         for creditnoteitem_data in creditnoteitems_data:
@@ -266,6 +328,11 @@ class CreditNoteSerializer(DocumentSerializer):
         return credit_note
 
     def update(self, instance, validated_data):
+        validated_data = {
+            **validated_data,
+            **self._get_calculated_fields(validated_data),
+        }
+
         creditnoteitems_data = validated_data.pop("creditnoteitem_set", [])
         customer = validated_data.get("customer")
         self._update_destroy_or_create(instance, creditnoteitems_data)
@@ -463,7 +530,106 @@ class InvoiceSerializer(DocumentSerializer):
         ).delete()
         InvoiceItem.objects.bulk_create(bulk_creates)
 
+    def _get_calculated_fields(self, validated_data):
+
+        discount_rate = validated_data.pop("discount_rate")
+        gst_rate = validated_data.pop("gst_rate")
+        invoiceitem_set = validated_data.pop("invoiceitem_set")
+        customer = validated_data.get("customer")
+        creditsapplication_set = validated_data.pop(
+            "creditsapplication_set", []
+        )
+        # TODO: fix round down behavior
+        # https://stackoverflow.com/questions/20457038/how-to-round-to-2-decimals-with-python
+        # round quantity, unit_price, gst_rate and discount rate first
+        # then round the rest at the end of calculation
+        # TODO: assign invoice later after create or update
+
+        credits_applied = Decimal("0.00")
+        if self.context["request"].method in ["PUT", "PATCH"]:
+            credits_applied = self.instance.credits_applied
+        new_creditsapplication_set = []
+        for credits_application in creditsapplication_set:
+            # update customer unused credits
+            amount_to_credit = round(
+                credits_application.get("amount_to_credit") or Decimal("0.00"),
+                2,
+            )
+            credit_note = credits_application.get("credit_note")
+
+            # next time when come to this, credits remaining will reduce
+            new_amount_to_credit = min(
+                amount_to_credit,
+                credit_note.credits_remaining,
+            )
+
+            credits_applied += new_amount_to_credit
+            credit_note.credits_remaining -= new_amount_to_credit
+            credit_note.credits_used += new_amount_to_credit
+            credit_note.save()
+
+            customer.unused_credits -= new_amount_to_credit
+            customer.save()
+
+            new_creditsapplication_set.append(
+                {
+                    **credits_application,
+                    "date": formats.date_format(datetime.now(), "Y-m-d"),
+                    "amount_to_credit": new_amount_to_credit,
+                }
+            )
+
+        invoiceitem_set = [
+            {
+                **invoiceitem,
+                "amount": round(
+                    round(invoiceitem.get("quantity"))
+                    * round(invoiceitem.get("unit_price"), 2),
+                    2,
+                ),
+            }
+            for invoiceitem in invoiceitem_set
+        ]
+        total_amount = sum(
+            [
+                # recalculate amount here since it has been rounded up
+                round(invoiceitem.get("quantity"))
+                * round(invoiceitem.get("unit_price"), 2)
+                for invoiceitem in invoiceitem_set
+            ]
+        )
+        discount_rate = round(discount_rate, 2)
+        gst_rate = round(gst_rate, 2)
+        discount_amount = total_amount * discount_rate / 100
+        net = total_amount * (1 - discount_rate / 100)
+        gst_amount = net * gst_rate / 100
+        grand_total = net * (1 + gst_rate / 100)
+        balance_due = grand_total - credits_applied
+
+        if balance_due < 0:
+            msg = _("Credits Applied cannot be more than Grand Total")
+            raise serializers.ValidationError(msg)
+
+        return {
+            "total_amount": round(total_amount, 2),
+            "discount_rate": discount_rate,
+            "gst_rate": gst_rate,
+            "discount_amount": round(discount_amount, 2),
+            "gst_amount": round(gst_amount, 2),
+            "net": round(net, 2),
+            "grand_total": round(grand_total, 2),
+            "invoiceitem_set": invoiceitem_set,
+            "creditsapplication_set": new_creditsapplication_set,
+            "credits_applied": credits_applied,
+            "balance_due": balance_due,
+        }
+
     def create(self, validated_data):
+        validated_data = {
+            **validated_data,
+            **self._get_calculated_fields(validated_data),
+        }
+
         invoiceitems_data = validated_data.pop("invoiceitem_set", [])
         creditsapplications_data = validated_data.pop(
             "creditsapplication_set", []
@@ -481,6 +647,11 @@ class InvoiceSerializer(DocumentSerializer):
         return invoice
 
     def update(self, instance, validated_data):
+        validated_data = {
+            **validated_data,
+            **self._get_calculated_fields(validated_data),
+        }
+
         invoiceitems_data = validated_data.pop("invoiceitem_set", [])
         creditsapplications_data = validated_data.pop(
             "creditsapplication_set", []
@@ -625,7 +796,57 @@ class SalesOrderSerializer(DocumentSerializer):
         ).delete()
         SalesOrderItem.objects.bulk_create(bulk_creates)
 
+    def _get_calculated_fields(self, validated_data):
+        discount_rate = validated_data.pop("discount_rate")
+        gst_rate = validated_data.pop("gst_rate")
+        salesorderitem_set = validated_data.pop("salesorderitem_set")
+        # TODO: fix round down behavior
+        # https://stackoverflow.com/questions/20457038/how-to-round-to-2-decimals-with-python
+        # round quantity, unit_price, gst_rate and discount rate first
+        # then round the rest at the end of calculation
+        salesorderitem_set = [
+            {
+                **salesorderitem,
+                "amount": round(
+                    round(salesorderitem.get("quantity"))
+                    * round(salesorderitem.get("unit_price"), 2),
+                    2,
+                ),
+            }
+            for salesorderitem in salesorderitem_set
+        ]
+        total_amount = sum(
+            [
+                # recalculate amount here since it has been rounded up
+                round(salesorderitem.get("quantity"))
+                * round(salesorderitem.get("unit_price"), 2)
+                for salesorderitem in salesorderitem_set
+            ]
+        )
+        discount_rate = round(discount_rate, 2)
+        gst_rate = round(gst_rate, 2)
+        discount_amount = total_amount * discount_rate / 100
+        net = total_amount * (1 - discount_rate / 100)
+        gst_amount = net * gst_rate / 100
+        grand_total = net * (1 + gst_rate / 100)
+
+        return {
+            "total_amount": round(total_amount, 2),
+            "discount_rate": discount_rate,
+            "gst_rate": gst_rate,
+            "discount_amount": round(discount_amount, 2),
+            "gst_amount": round(gst_amount, 2),
+            "net": round(net, 2),
+            "grand_total": round(grand_total, 2),
+            "salesorderitem_set": salesorderitem_set,
+        }
+
     def create(self, validated_data):
+        validated_data = {
+            **validated_data,
+            **self._get_calculated_fields(validated_data),
+        }
+
         salesorderitems_data = validated_data.pop("salesorderitem_set", [])
         sales_order = SalesOrder.objects.create(**validated_data)
         for salesorderitem_data in salesorderitems_data:
@@ -643,6 +864,11 @@ class SalesOrderSerializer(DocumentSerializer):
         return sales_order
 
     def update(self, instance, validated_data):
+        validated_data = {
+            **validated_data,
+            **self._get_calculated_fields(validated_data),
+        }
+
         salesorderitems_data = validated_data.pop("salesorderitem_set", [])
 
         self._update_destroy_or_create(instance, salesorderitems_data)
