+++ customer/views.py
@@ -1,19 +1,14 @@
 from core.utils import validate_reference_uniqueness
-from decimal import Decimal
-from datetime import datetime
 
-from django.utils import formats
 from django.utils.translation import ugettext_lazy as _
 
 from django_filters import rest_framework as filters
-from rest_framework import viewsets, mixins, serializers
+from rest_framework import viewsets, mixins
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
 
-from core.utils import validate_reference_uniqueness_in_data
-
-
-from core.views import BaseAssetAttrViewSet
+from core.utils import validate_bulk_reference_uniqueness
+from core.views import BaseAssetAttrViewSet, BaseDocumentViewSet
 from core.models import (
     Invoice,
     Customer,
@@ -56,13 +51,13 @@ class CustomerViewSet(BaseAssetAttrViewSet):
         "receivables",
     ]
 
-    def perform_create(self, serializer):
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        super().perform_create(serializer)
+    def perform_bulk_create(self, serializer):
+        validate_bulk_reference_uniqueness(serializer.validated_data)
+        return self.perform_create(serializer)
 
-    def perform_update(self, serializer):
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        serializer.save()
+    def perform_bulk_update(self, serializer):
+        validate_bulk_reference_uniqueness(serializer.validated_data)
+        return self.perform_update(serializer)
 
 
 class CreditsApplicationFilter(filters.FilterSet):
@@ -130,7 +125,7 @@ class CreditNoteFilter(filters.FilterSet):
         }
 
 
-class CreditNoteViewSet(BaseAssetAttrViewSet):
+class CreditNoteViewSet(BaseDocumentViewSet):
     """Manage credit note in the database"""
 
     queryset = CreditNote.objects.all()
@@ -146,74 +141,6 @@ class CreditNoteViewSet(BaseAssetAttrViewSet):
         "grand_total",
     ]
 
-    def _get_calculated_fields(self, serializer):
-        # assume all positive
-        discount_rate = serializer.validated_data.pop("discount_rate")
-        gst_rate = serializer.validated_data.pop("gst_rate")
-        creditnoteitem_set = serializer.validated_data.pop(
-            "creditnoteitem_set"
-        )
-        credits_used = Decimal("0.00")
-        # TODO: remove this later after implement apply credits to invoice from credit note
-        if self.request.method in ["PUT", "PATCH"]:
-            credits_used = serializer.instance.credits_used
-
-        refund = serializer.validated_data.pop("refund")
-
-        creditnoteitem_set = [
-            {
-                **creditnoteitem,
-                "amount": round(
-                    round(creditnoteitem.get("quantity"))
-                    * round(creditnoteitem.get("unit_price"), 2),
-                    2,
-                ),
-            }
-            for creditnoteitem in creditnoteitem_set
-        ]
-        total_amount = sum(
-            [
-                # recalculate amount here since it has been rounded up
-                round(creditnoteitem.get("quantity"))
-                * round(creditnoteitem.get("unit_price"), 2)
-                for creditnoteitem in creditnoteitem_set
-            ]
-        )
-        discount_rate = round(discount_rate, 2)
-        gst_rate = round(gst_rate, 2)
-        discount_amount = total_amount * discount_rate / 100
-        net = total_amount * (1 - discount_rate / 100)
-        gst_amount = net * gst_rate / 100
-        grand_total = net * (1 + gst_rate / 100)
-        credits_remaining = grand_total - credits_used
-        refund = min(credits_remaining, round(refund, 2))
-        credits_remaining -= refund
-
-        return {
-            "total_amount": round(total_amount, 2),
-            "discount_rate": discount_rate,
-            "gst_rate": gst_rate,
-            "discount_amount": round(discount_amount, 2),
-            "gst_amount": round(gst_amount, 2),
-            "net": round(net, 2),
-            "grand_total": round(grand_total, 2),
-            "creditnoteitem_set": creditnoteitem_set,
-            "credits_used": credits_used,
-            "credits_remaining": credits_remaining,
-            "refund": refund,
-        }
-
-    def perform_create(self, serializer):
-        company = self.request.user.company
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        serializer.save(
-            company=company, **self._get_calculated_fields(serializer)
-        )
-
-    def perform_update(self, serializer):
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        serializer.save(**self._get_calculated_fields(serializer))
-
 
 class InvoiceFilter(filters.FilterSet):
     class Meta:
@@ -228,7 +155,7 @@ class InvoiceFilter(filters.FilterSet):
         }
 
 
-class InvoiceViewSet(BaseAssetAttrViewSet):
+class InvoiceViewSet(BaseDocumentViewSet):
     """Manage invoice in the database"""
 
     queryset = Invoice.objects.all()
@@ -244,110 +171,6 @@ class InvoiceViewSet(BaseAssetAttrViewSet):
         "grand_total",
     ]
 
-    def _get_calculated_fields(self, serializer):
-        discount_rate = serializer.validated_data.pop("discount_rate")
-        gst_rate = serializer.validated_data.pop("gst_rate")
-        invoiceitem_set = serializer.validated_data.pop("invoiceitem_set")
-        customer = serializer.validated_data.get("customer")
-        creditsapplication_set = serializer.validated_data.pop(
-            "creditsapplication_set", []
-        )
-        # TODO: fix round down behavior
-        # https://stackoverflow.com/questions/20457038/how-to-round-to-2-decimals-with-python
-        # round quantity, unit_price, gst_rate and discount rate first
-        # then round the rest at the end of calculation
-        # TODO: assign invoice later after create or update
-
-        credits_applied = Decimal("0.00")
-        if self.request.method in ["PUT", "PATCH"]:
-            credits_applied = serializer.instance.credits_applied
-        new_creditsapplication_set = []
-        for credits_application in creditsapplication_set:
-            # update customer unused credits
-            amount_to_credit = round(
-                credits_application.get("amount_to_credit") or Decimal("0.00"),
-                2,
-            )
-            credit_note = credits_application.get("credit_note")
-
-            # next time when come to this, credits remaining will reduce
-            new_amount_to_credit = min(
-                amount_to_credit,
-                credit_note.credits_remaining,
-            )
-
-            credits_applied += new_amount_to_credit
-            credit_note.credits_remaining -= new_amount_to_credit
-            credit_note.credits_used += new_amount_to_credit
-            credit_note.save()
-
-            customer.unused_credits -= new_amount_to_credit
-            customer.save()
-
-            new_creditsapplication_set.append(
-                {
-                    **credits_application,
-                    "date": formats.date_format(datetime.now(), "Y-m-d"),
-                    "amount_to_credit": new_amount_to_credit,
-                }
-            )
-
-        invoiceitem_set = [
-            {
-                **invoiceitem,
-                "amount": round(
-                    round(invoiceitem.get("quantity"))
-                    * round(invoiceitem.get("unit_price"), 2),
-                    2,
-                ),
-            }
-            for invoiceitem in invoiceitem_set
-        ]
-        total_amount = sum(
-            [
-                # recalculate amount here since it has been rounded up
-                round(invoiceitem.get("quantity"))
-                * round(invoiceitem.get("unit_price"), 2)
-                for invoiceitem in invoiceitem_set
-            ]
-        )
-        discount_rate = round(discount_rate, 2)
-        gst_rate = round(gst_rate, 2)
-        discount_amount = total_amount * discount_rate / 100
-        net = total_amount * (1 - discount_rate / 100)
-        gst_amount = net * gst_rate / 100
-        grand_total = net * (1 + gst_rate / 100)
-        balance_due = grand_total - credits_applied
-
-        if balance_due < 0:
-            msg = _("Credits Applied cannot be more than Grand Total")
-            raise serializers.ValidationError(msg)
-
-        return {
-            "total_amount": round(total_amount, 2),
-            "discount_rate": discount_rate,
-            "gst_rate": gst_rate,
-            "discount_amount": round(discount_amount, 2),
-            "gst_amount": round(gst_amount, 2),
-            "net": round(net, 2),
-            "grand_total": round(grand_total, 2),
-            "invoiceitem_set": invoiceitem_set,
-            "creditsapplication_set": new_creditsapplication_set,
-            "credits_applied": credits_applied,
-            "balance_due": balance_due,
-        }
-
-    def perform_create(self, serializer):
-        company = self.request.user.company
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        serializer.save(
-            company=company, **self._get_calculated_fields(serializer)
-        )
-
-    def perform_update(self, serializer):
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        serializer.save(**self._get_calculated_fields(serializer))
-
 
 class SalesOrderFilter(filters.FilterSet):
     class Meta:
@@ -357,7 +180,7 @@ class SalesOrderFilter(filters.FilterSet):
         }
 
 
-class SalesOrderViewSet(BaseAssetAttrViewSet):
+class SalesOrderViewSet(BaseDocumentViewSet):
     """Manage customer in the database"""
 
     queryset = SalesOrder.objects.all()
@@ -371,61 +194,3 @@ class SalesOrderViewSet(BaseAssetAttrViewSet):
         "status",
         "grand_total",
     ]
-
-    def _get_calculated_fields(self, serializer):
-        discount_rate = serializer.validated_data.pop("discount_rate")
-        gst_rate = serializer.validated_data.pop("gst_rate")
-        salesorderitem_set = serializer.validated_data.pop(
-            "salesorderitem_set"
-        )
-        # TODO: fix round down behavior
-        # https://stackoverflow.com/questions/20457038/how-to-round-to-2-decimals-with-python
-        # round quantity, unit_price, gst_rate and discount rate first
-        # then round the rest at the end of calculation
-        salesorderitem_set = [
-            {
-                **salesorderitem,
-                "amount": round(
-                    round(salesorderitem.get("quantity"))
-                    * round(salesorderitem.get("unit_price"), 2),
-                    2,
-                ),
-            }
-            for salesorderitem in salesorderitem_set
-        ]
-        total_amount = sum(
-            [
-                # recalculate amount here since it has been rounded up
-                round(salesorderitem.get("quantity"))
-                * round(salesorderitem.get("unit_price"), 2)
-                for salesorderitem in salesorderitem_set
-            ]
-        )
-        discount_rate = round(discount_rate, 2)
-        gst_rate = round(gst_rate, 2)
-        discount_amount = total_amount * discount_rate / 100
-        net = total_amount * (1 - discount_rate / 100)
-        gst_amount = net * gst_rate / 100
-        grand_total = net * (1 + gst_rate / 100)
-
-        return {
-            "total_amount": round(total_amount, 2),
-            "discount_rate": discount_rate,
-            "gst_rate": gst_rate,
-            "discount_amount": round(discount_amount, 2),
-            "gst_amount": round(gst_amount, 2),
-            "net": round(net, 2),
-            "grand_total": round(grand_total, 2),
-            "salesorderitem_set": salesorderitem_set,
-        }
-
-    def perform_create(self, serializer):
-        company = self.request.user.company
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        serializer.save(
-            company=company, **self._get_calculated_fields(serializer)
-        )
-
-    def perform_update(self, serializer):
-        validate_reference_uniqueness_in_data(serializer.validated_data)
-        serializer.save(**self._get_calculated_fields(serializer))
